requirements.txt
mysql-connector-python==8.0.31
flask==2.1.0
pandas>=1.5.0
rfeed>=1.1.1